Move non-page components to /components directory
