+++ components/activity/Resource.js
